Add main functionality
Add ClientList. This class is used to store all writers for all clients and also, to store all client's inetAdresses.
Add EchoThread. This thread is used for communication with connected user. It can receive message from client and send it to all other clients.
Add execute method to Server class to execute EchoThread when client is connected.